removes file renaming logic from action generation file and moves action test templates into a __tests__ folder
