+++ packages/cli/templates/actions/empty-action/__tests__/index.test.ts
@@ -1,6 +1,6 @@
 import nock from 'nock'
 import { createTestEvent, createTestIntegration } from '@segment/actions-core'
-import Destination from '../index'
+import Destination from '../../index'
 
 const testDestination = createTestIntegration(Destination)
