+++ packages/cli/templates/actions/empty-browser-action/__tests__/index.test.ts
@@ -1,6 +1,6 @@
 import { Analytics, Context } from '@segment/analytics-next'
 import * as jsdom from 'jsdom'
-import { destination } from '../index'
+import { destination } from '../../index'
 
 describe('{{destination}}', () => {
   describe('{{name}}', () => {
